Import and view Twilio call data
This only saves valid Twilio calls, where a valid
call is one that has not an inbound and outbound
part and the outbound part has a duration > 10sec.

As a result this code ignores any calls that exist
of an inbound part only are logged and then dropped
from the processing stream.