Add simple import for google-based data
Imports the google forms backed 'where did you hear' data. This is a
one-off task that can be scrapped after the cut-off period for our
usage of google forms for the 'where did you hear' data.

At the moment we just take the data as-is. Some of this appears to be
of 'varying' quality. I'm sure we'll need some manual cleanup once
complete.