Add Mail Retriever to abstract IMAP interface
This class is designed to be used to retrieve
attachments from emails.

Emails can be selected by filter keys and then
the attachment selected using a name RegExp.

Emails should be archived post-processing using
the 'archive' method.

Care should be taken to handle any failure that
occurs before the email is moved and after it has
been processed. Ideally all imports that use this
code should be Idempotent.