Ensure HTML5 date pickers can be bound
Prior to this change, `#to_s` on date-like objects would include the
time portion, thus causing warnings and unexpected results when
attempting to bind the value to an HTML5 date picker. This change
ensures we strip the time portion when presenting a `DateTime`.

Traditionally, I'd prefer an integration spec that proves this
behaviour, but capybara has problems with HTML5 date pickers so that
proved fruitless.