Import TPAS Satisfaction data
Raw(ish) data will be provided via daily email.

Matching row are grouped and then assigned an integer
indicating which 'row' they represent, this data is then
used to create a UID, allowing data to be imported
multiple times.

NOTE: This process only supports growing datasets,
if the number of rows decreases, this will NOT be
correctly updated.