Fix bug with costs breakdown report when no twilio calls
This bug presents when entering data for May 2015 as
twilio was only used from Aug 2015 onwards.

This fix uses calls volumes from the next month with any
calls if no calls exists for the given month.

An additionally fallback has been added in case future cost
data is entered. In this case it will split based on
call volumes from the previous month with call data.