Accommodate new exit poll for reporting
The exit polls have changed for face-to-face appointments and we need
to alter the way we process and import the google satisfaction data to
match the new shape and structure of this data.

The headline changes are:

* Keyword changes for the raw satisfaction value
* Simplified locations, raw satisfaction (unified positionally)
* Accurate appointment dates

Appointment times are fudged to `09:00` since the time portion of the
appointment is not accurately reported. Previously we were incorrectly
using the date/time from the `timestamp`. This represents the date/time
the data was entered, not the actual appointment date/time.